backend-drf/requirements.txt
tensorflow==2.15.0
torch==2.2.0
scikit-learn==1.4.0
numpy==1.26.0
django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8